made a separate file for the appdata object. The appdata object is no longer a global variable, but is passed as arguments to progressively lower-level widgets. Fixes #17
