+++ src/lib/main.dart
@@ -9,125 +9,61 @@ import 'package:uuid/uuid.dart';
 import 'fs.dart';
 import 'util.dart';
 import 'shop-list-entry.dart';
+import 'models/appdata.dart';
 
-void main() => runApp(MyApp());
+Future<void> main() async {
 
-// TODO: calling load works fine, but the widgets onscreen don't update until some action is performed. boo.
+  WidgetsFlutterBinding.ensureInitialized();
 
-class AppData{
-    /// All information that must persist between closing and opening the app
-    ///
-    /// The AppData object can be mapped to json-formatted string and stored on the device.
-    /// It can be restored from such a file on the device as well.
-    ///
-    /// There's two major attributes; homeList and supermarketOrder
-    /// The homeList is the shopping list the user crafts at home.
-    ///   It consists of all the products he or she might need.
-    ///   The user can toggle each product to cross it off the list (without hard-removing it)
-    ///   The user can add structure by adding headers ("this is what I need for tuesday", "this is what I need for that pie i want to bake"), purely for bookkeeping
-    /// The supermarketOrder is the order in which the products appear if the user follows his or her favourite route
-    ///   We're not going to attempt to find the shortest route through the supermarket based on product locations
-    ///   That's a travelling salesman problem, which is out of scope for this simple app
-    ///   We assume the user always walks the same route through the supermarket regardless of what items he or she needs
-    ///
-    /// Each entry in the homeList has a uuid (universally unique identifier) because it needs that to be reorderable
-    /// The name is not the uuid, because the user might enter 'tomatoes' twice, for example
-    /// In the supermarketOrder the name can be the uuid, because we won't allow adding the same entry twice there
+  AppData appData = AppData();
 
-    // TODO: Can't move AppData to a different file because it will throw errors like 'appdata.store()' is not defined and such
+  runApp(MainApp(appData));
 
-    static final AppData _appData = new AppData._internal();
-
-    List homeList = [
-        HeaderEntry(id: Uuid().v1(), text: "Monday"),
-        ProductEntry(id: Uuid().v1(), text: "pasta"),
-        ProductEntry(id: Uuid().v1(), text: "tomatoes"),
-        HeaderEntry(id: Uuid().v1(), text: "Tuesday"),
-        ProductEntry(id: Uuid().v1(), text: "rice"),
-        ProductEntry(id: Uuid().v1(), text: "broccoli"),
-        HeaderInputField(id: Uuid().v1()),
-        ProductInputField(id: Uuid().v1()),
-    ];
-
-    List supermarketOrder = [
-      "bananas",
-      "tomatoes",
-      "broccoli",
-      "rice",
-      "pasta",
-      "ice cream",
-      ""
-    ];
-
-    void fromJson(Map<String, dynamic> json) {
-      homeList = json["homeList"].map<Entry>((x) => Entry.fromJson(x)).toList();
-      supermarketOrder = json["supermarketOrder"];
-    }
-
-    Map<String, dynamic> toJson() {
-      return {
-        "homeList": homeList.map((x) => x.toJson()).toList(),
-        "supermarketOrder": supermarketOrder,
-      };
-    }
-
-    Future<void> store() async {
-        Map<String, dynamic> json = toJson();
-        // jsonPrettyPrint(json); // for debugging
-        writeAppDataString(jsonEncode(json));
-        print("Done writing appData to disk");
-    }
-
-    Future<void> load() async {
-        return loadAppDataString().then((String jsonString) {
-            Map<String, dynamic> json = jsonDecode(jsonString);
-            // jsonPrettyPrint(json); // for debugging
-            fromJson(json);
-            print("Done loading appData from disk");
-        });
-    }
-
-    factory AppData() {
-        return _appData;
-    }
-
-    AppData._internal();
 }
 
-AppData appData = AppData();
+class MainApp extends StatelessWidget {
 
-// the MAIN app (stateless)
-class MyApp extends StatelessWidget {
+    AppData appData;
+
+    MainApp(this.appData);
 
     @override
     Widget build(BuildContext context) {
         return MaterialApp(
             title : 'ShopList',
-            home  : BottomBarMainBodyWidget()
+            home  : BottomBarMainBodyWidget(appData)
         );
     }
 }
 
 // Show stuff that's always there; the bottom bar and an open window where the main window will go
 class BottomBarMainBodyWidget extends StatefulWidget {
-    BottomBarMainBodyWidget({Key key}) : super(key: key);
+
+    AppData appData;
+
+    BottomBarMainBodyWidget(this.appData);
 
     @override
-    _BottomBarMainBodyWidgetState createState() => _BottomBarMainBodyWidgetState();
+    _BottomBarMainBodyWidgetState createState() => _BottomBarMainBodyWidgetState(appData);
 }
 
 class _BottomBarMainBodyWidgetState extends State<BottomBarMainBodyWidget> {
     /// Holds a list of tabs and a bottom bar to let the user switch between tabs
 
+    AppData appData;
+
+    _BottomBarMainBodyWidgetState(this.appData);
+
     int _selectedPageIndex = 0;
-    List<Widget> _widgetOptions = <Widget>[
-        HomeList(),
-        TripList(),
-        SupermarketList()
-    ];
 
     @override
     Widget build(BuildContext context) {
+
+        List<Widget> _widgetOptions = <Widget>[
+            HomeList(appData),
+            TripList(appData),
+            SupermarketList(appData)
+        ];
 
         return Scaffold(
             body: Center(
@@ -165,10 +101,12 @@ class _BottomBarMainBodyWidgetState extends State<BottomBarMainBodyWidget> {
 //
 class HomeList extends StatefulWidget {
 
-    HomeList({ Key key}): super(key: key);
+    AppData appData;
+
+    HomeList(this.appData);
 
     @override
-    _HomeListState createState() => _HomeListState();
+    _HomeListState createState() => _HomeListState(appData);
 }
 
 class _HomeListState extends State<HomeList> {
@@ -184,6 +122,10 @@ class _HomeListState extends State<HomeList> {
     /// The user can drag and reorder existing entries
     /// The user can remove ProductEntries and HeaderEntries
     /// The user cannot remove ProductEntryFields and HeaderEntryFields
+
+    AppData appData;
+
+    _HomeListState(this.appData);
 
     @override
     void initState() {
@@ -451,10 +393,12 @@ class _HomeListState extends State<HomeList> {
 //
 class TripList extends StatefulWidget {
 
-    TripList({ Key key}): super(key: key);
+    AppData appData;
+
+    TripList(this.appData);
 
     @override
-    _TripListState createState() => _TripListState();
+    _TripListState createState() => _TripListState(appData);
 }
 
 class _TripListState extends State<TripList> {
@@ -470,6 +414,10 @@ class _TripListState extends State<TripList> {
     ///   There was some logic for this in an older version of the app, see if mimicking that is appropriate
     ///
     /// TODO: allow ProductEntryField here, new entry should be added to homeList in a trivial position (bottom?)
+
+    AppData appData;
+
+    _TripListState(this.appData);
 
     @override
     void initState() {
@@ -556,10 +504,12 @@ class _TripListState extends State<TripList> {
 //
 class SupermarketList extends StatefulWidget {
 
-    SupermarketList({ Key key}): super(key: key);
+    AppData appData;
+
+    SupermarketList(this.appData);
 
     @override
-    _SupermarketListState createState() => _SupermarketListState();
+    _SupermarketListState createState() => _SupermarketListState(appData);
 }
 
 class _SupermarketListState extends State<SupermarketList> {
@@ -571,6 +521,10 @@ class _SupermarketListState extends State<SupermarketList> {
     ///
     /// TODO: the supermarketOrder is a list of strings whereas the homeList is a list of Entry objects, it'd be prettier if supermarketOrder is also a list of Entries (subclasses like SupermarketOrderProductEntry?)
     /// TODO: allow multiple customizable supermarketorderings (MyLocalJumbo, ThatOneBigAlbertHeijn, etc...)
+
+    AppData appData;
+
+    _SupermarketListState(this.appData);
 
     @override
     void initState() {
